requirements.txt
pandas>=2.0
selenium>=4.10
undetected-chromedriver>=3.4.0
python-dotenv>=1.0
sqlalchemy>=2.0
pyodbc>=4.0
jupyter>=1.0
requests>=2.31
beautifulsoup4>=4.12
lxml>=4.9